requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==1.4.46
alembic==1.9.2
asyncpg==0.27.0
pydantic[email]~=1.10.4
alembic
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
python-multipart
numpy~=1.24.2
Pillow~=9.4.0